script: add flag to not build react components

+++ script/build.sh
@@ -12,26 +12,28 @@ echo "✅ Unzip completed"
 npm run svgo
 echo "✅ Icons optimized"
 
-mkdir 'temp-icons'
-echo "✅ temp-icons direcotry created"
+if [[ $* != *-s* ]]; then
+  mkdir 'temp-icons'
+  echo "✅ temp-icons direcotry created"
 
-npm run build:temp
-echo "✅ Icons copied to a new directory"
+  npm run build:temp
+  echo "✅ Icons copied to a new directory"
 
-npm run build:icons
-echo "✅ React component files created"
+  npm run build:icons
+  echo "✅ React component files created"
 
-npm run build:data
-# Success messages are inside the script
+  npm run build:data
+  # Success messages are inside the script
 
-npm run babel
-echo "✅ Icons correctly compiled"
+  npm run babel
+  echo "✅ Icons correctly compiled"
 
-npm run test
-echo "✅ Tests are successfull"
+  npm run test
+  echo "✅ Tests are successfull"
 
-rm -rf 'temp-icons'
-echo "✅ temp-icons directory removed"
+  rm -rf 'temp-icons'
+  echo "✅ temp-icons directory removed"
+fi
 
 echo ""
 echo "✨  Success"
